+++ src/Components/ChartsSection/CombinedCharts.jsx
@@ -1,13 +1,13 @@
+import { Flex } from "@chakra-ui/react";
 import React from "react";
 import LineChart from "./LineChart";
 import PieChart from "./PieChart";
-import { Flex } from "@chakra-ui/react";
 
 const CombinedCharts = () => {
   return (
     <Flex
       w="full"
-      direction="row"
+      direction={{ base: "column", lg: "row" }}
       align={"center"}
       justifyContent="space-between"
       // px="2rem"
